# Implement SIMD versions of vector operations