## Rationale

Binary distributions of `pgvector` often need to disable the `-march=native` flag (e.g., [Debian patch](https://salsa.debian.org/postgresql/pgvector/-/blob/develop/debian/patches/no-native)) because, during build time, one cannot know in advance where the binary will be running. As a result, loops inside the vector operations cannot be auto-vectorized (contrary to the assumption in the code), which results in mediocre performances in places where the binary distributions are used (e.g., cloud providers).

To alleviate the above, I have implemented the SSE, AVX and AVX512F versions of the vector operations, and added a CPU dispatching mechanism (during extension load time) to pick the most recent version the underlying CPU supports (by following the best practices mentioned in Agner Fog's [Optimizing Software in C++ (Chapters 12 and 13)](https://www.agner.org/optimize/#manuals) manual).

## Benchmark

I have created a separate [`simd-playground`](https://github.com/aytekinar/simd-playground) repository where I host the different SIMD implementations of the relevant vector operations, together with their unit tests and benchmarks.

When doing the benchmarks, I have used the following flags:

- GCC/Clang (non-native): `-Wall -Wpedantic -O2 -DNDEBUG -ftree-vectorize -fassociative-math -fno-trapping-math -fno-math-errno -fno-signed-zeros -funroll-loops`
- GCC/Clang (native): `-Wall -Wpedantic -O2 -DNDEBUG -ftree-vectorize -fassociative-math -fno-trapping-math -fno-math-errno -fno-signed-zeros -funroll-loops -march=native`
- MSVC (non-native): `/Wall /O2 /fp:fast`
- MSVC (native): `/Wall /O2 /fp:fast /arch:AVX512`

### C Benchmark

Below, I provide the benchmark results on my machine (with 11th Gen Intel(R) Core(TM) i9-11900H @ 2.50GHz). Note that the benchmark name is in the form `BM_<vector_op>/{1,2,3,4}/<vector_len>`, where `1: scalar`, `2: sse`, `3: avx` and `4: avx512f`. The scalar version is simply a for-loop, which is vectorized in `native`-builds, and the rest are manually implemented vectorized versions for the respective targets:

<details><summary>Non-native builds</summary>

| Benchmark                  | Time (MSVC) | CPU (MSVC) | Iterations (MSVC) | Time (GCC) | CPU (GCC) | Iterations (GCC) | Time (Clang) | CPU (Clang) | Iterations (Clang) |
| -------------------------- | ----------- | ---------- | ----------------- | ---------- | --------- | ---------------- | ------------ | ----------- | ------------------ |
| BM_add/1/1024              | 128 ns      | 126 ns     | 5600000           | 90.6 ns    | 90.6 ns   | 7677939          | 124 ns       | 124 ns      | 5583785            |
| BM_add/2/1024              | 83 ns       | 83.7 ns    | 7466667           | 89.7 ns    | 89.7 ns   | 7573045          | 134 ns       | 134 ns      | 5201665            |
| BM_add/3/1024              | 46.9 ns     | 47.1 ns    | 14933333          | 51.1 ns    | 51 ns     | 13665016         | 52.6 ns      | 52.6 ns     | 13217498           |
| BM_add/4/1024              | 39.7 ns     | 38.1 ns    | 17230769          | 41.8 ns    | 41.8 ns   | 17133409         | 37.6 ns      | 37.6 ns     | 18376567           |
| BM_add/1/2048              | 204 ns      | 199 ns     | 3446154           | 175 ns     | 175 ns    | 4002587          | 243 ns       | 243 ns      | 2799188            |
| BM_add/2/2048              | 220 ns      | 220 ns     | 3200000           | 180 ns     | 180 ns    | 3895405          | 262 ns       | 262 ns      | 2545064            |
| BM_add/3/2048              | 111 ns      | 112 ns     | 6400000           | 98 ns      | 98 ns     | 7077248          | 109 ns       | 109 ns      | 6344137            |
| BM_add/4/2048              | 114 ns      | 109 ns     | 5600000           | 78.9 ns    | 78.9 ns   | 8771963          | 76.1 ns      | 76.1 ns     | 9199390            |
| BM_add/1/4096              | 456 ns      | 449 ns     | 1600000           | 370 ns     | 370 ns    | 1897552          | 504 ns       | 504 ns      | 1360571            |
| BM_add/2/4096              | 516 ns      | 531 ns     | 1000000           | 373 ns     | 373 ns    | 1889332          | 546 ns       | 546 ns      | 1204813            |
| BM_add/3/4096              | 265 ns      | 261 ns     | 2635294           | 209 ns     | 209 ns    | 3339430          | 227 ns       | 227 ns      | 2922807            |
| BM_add/4/4096              | 151 ns      | 146 ns     | 4480000           | 168 ns     | 168 ns    | 4184801          | 162 ns       | 162 ns      | 4124729            |
| BM_add/1/8192              | 1177 ns     | 1172 ns    | 560000            | 1035 ns    | 1035 ns   | 669624           | 1088 ns      | 1088 ns     | 627519             |
| BM_add/2/8192              | 1375 ns     | 1245 ns    | 640000            | 1077 ns    | 1077 ns   | 661930           | 1230 ns      | 1230 ns     | 539861             |
| BM_add/3/8192              | 822 ns      | 795 ns     | 746667            | 737 ns     | 737 ns    | 935601           | 832 ns       | 832 ns      | 840245             |
| BM_add/4/8192              | 983 ns      | 977 ns     | 896000            | 763 ns     | 763 ns    | 919993           | 769 ns       | 769 ns      | 838225             |
| BM_add/1/16384             | 3214 ns     | 3181 ns    | 235789            | 1943 ns    | 1943 ns   | 361022           | 2162 ns      | 2162 ns     | 317620             |
| BM_add/2/16384             | 3164 ns     | 3069 ns    | 224000            | 1972 ns    | 1972 ns   | 357117           | 2467 ns      | 2467 ns     | 274078             |
| BM_add/3/16384             | 1834 ns     | 1814 ns    | 344615            | 1463 ns    | 1463 ns   | 459819           | 1644 ns      | 1644 ns     | 418869             |
| BM_add/4/16384             | 1620 ns     | 1573 ns    | 407273            | 1521 ns    | 1521 ns   | 461963           | 1548 ns      | 1548 ns     | 458229             |
| BM_sub/1/1024              | 147 ns      | 145 ns     | 5600000           | 87.7 ns    | 87.7 ns   | 7951808          | 124 ns       | 124 ns      | 5680735            |
| BM_sub/2/1024              | 139 ns      | 138 ns     | 4977778           | 94.7 ns    | 94.7 ns   | 7756491          | 133 ns       | 133 ns      | 5063914            |
| BM_sub/3/1024              | 60.5 ns     | 61.4 ns    | 11200000          | 51.2 ns    | 51.2 ns   | 11758678         | 52.7 ns      | 52.7 ns     | 13341999           |
| BM_sub/4/1024              | 93.8 ns     | 92.1 ns    | 7466667           | 41.9 ns    | 41.9 ns   | 16740797         | 38.1 ns      | 38.1 ns     | 18706934           |
| BM_sub/1/2048              | 274 ns      | 279 ns     | 2635294           | 175 ns     | 175 ns    | 4003315          | 245 ns       | 245 ns      | 2853324            |
| BM_sub/2/2048              | 290 ns      | 279 ns     | 2635294           | 181 ns     | 181 ns    | 3861775          | 267 ns       | 267 ns      | 2625535            |
| BM_sub/3/2048              | 120 ns      | 117 ns     | 5600000           | 99.5 ns    | 99.5 ns   | 7032434          | 107 ns       | 107 ns      | 6480215            |
| BM_sub/4/2048              | 183 ns      | 184 ns     | 4072727           | 79.6 ns    | 79.6 ns   | 8745474          | 76.1 ns      | 76.1 ns     | 8960849            |
| BM_sub/1/4096              | 462 ns      | 446 ns     | 1120000           | 367 ns     | 367 ns    | 1888530          | 501 ns       | 501 ns      | 1382872            |
| BM_sub/2/4096              | 469 ns      | 465 ns     | 1445161           | 372 ns     | 372 ns    | 1874261          | 545 ns       | 545 ns      | 1187354            |
| BM_sub/3/4096              | 260 ns      | 257 ns     | 2488889           | 210 ns     | 210 ns    | 3326323          | 226 ns       | 226 ns      | 3079691            |
| BM_sub/4/4096              | 158 ns      | 157 ns     | 4977778           | 168 ns     | 168 ns    | 4154352          | 162 ns       | 162 ns      | 4366883            |
| BM_sub/1/8192              | 1151 ns     | 1161 ns    | 497778            | 1044 ns    | 1044 ns   | 665623           | 1071 ns      | 1071 ns     | 645419             |
| BM_sub/2/8192              | 1171 ns     | 1172 ns    | 640000            | 1046 ns    | 1046 ns   | 662160           | 1238 ns      | 1238 ns     | 558128             |
| BM_sub/3/8192              | 780 ns      | 711 ns     | 746667            | 735 ns     | 735 ns    | 944414           | 823 ns       | 823 ns      | 789451             |
| BM_sub/4/8192              | 791 ns      | 802 ns     | 896000            | 756 ns     | 756 ns    | 873892           | 770 ns       | 770 ns      | 893439             |
| BM_sub/1/16384             | 2541 ns     | 2567 ns    | 280000            | 1938 ns    | 1938 ns   | 363417           | 2193 ns      | 2193 ns     | 265749             |
| BM_sub/2/16384             | 2512 ns     | 2448 ns    | 248889            | 2031 ns    | 2031 ns   | 332417           | 2443 ns      | 2443 ns     | 287533             |
| BM_sub/3/16384             | 1702 ns     | 1709 ns    | 448000            | 1475 ns    | 1475 ns   | 470820           | 1664 ns      | 1664 ns     | 401319             |
| BM_sub/4/16384             | 1700 ns     | 1674 ns    | 448000            | 1537 ns    | 1537 ns   | 455991           | 1531 ns      | 1531 ns     | 440341             |
| BM_dot_product/1/1024      | 111 ns      | 103 ns     | 6400000           | 188 ns     | 188 ns    | 3698332          | 102 ns       | 102 ns      | 6738661            |
| BM_dot_product/2/1024      | 210 ns      | 205 ns     | 3200000           | 188 ns     | 188 ns    | 3694425          | 146 ns       | 146 ns      | 4747852            |
| BM_dot_product/3/1024      | 84.2 ns     | 83.7 ns    | 8960000           | 79.8 ns    | 79.8 ns   | 8554048          | 66.2 ns      | 66.2 ns     | 10392464           |
| BM_dot_product/4/1024      | 49.1 ns     | 48 ns      | 16592593          | 38 ns      | 38 ns     | 18404189         | 38.4 ns      | 38.4 ns     | 18331705           |
| BM_dot_product/1/2048      | 217 ns      | 220 ns     | 3200000           | 404 ns     | 404 ns    | 1724995          | 213 ns       | 213 ns      | 3272856            |
| BM_dot_product/2/2048      | 423 ns      | 424 ns     | 1659259           | 404 ns     | 404 ns    | 1735376          | 311 ns       | 311 ns      | 2277208            |
| BM_dot_product/3/2048      | 189 ns      | 193 ns     | 3733333           | 186 ns     | 186 ns    | 3700616          | 147 ns       | 147 ns      | 4701893            |
| BM_dot_product/4/2048      | 96.3 ns     | 92.1 ns    | 5600000           | 79.5 ns    | 79.5 ns   | 8760611          | 80.8 ns      | 80.8 ns     | 8602193            |
| BM_dot_product/1/4096      | 462 ns      | 465 ns     | 1544828           | 836 ns     | 836 ns    | 834767           | 426 ns       | 426 ns      | 1636841            |
| BM_dot_product/2/4096      | 907 ns      | 879 ns     | 746667            | 834 ns     | 834 ns    | 818919           | 635 ns       | 635 ns      | 1083863            |
| BM_dot_product/3/4096      | 421 ns      | 417 ns     | 1723077           | 404 ns     | 404 ns    | 1726134          | 312 ns       | 312 ns      | 2163755            |
| BM_dot_product/4/4096      | 217 ns      | 222 ns     | 3733333           | 183 ns     | 183 ns    | 3793057          | 186 ns       | 186 ns      | 3734224            |
| BM_dot_product/1/8192      | 919 ns      | 921 ns     | 746667            | 1696 ns    | 1696 ns   | 411496           | 881 ns       | 881 ns      | 781927             |
| BM_dot_product/2/8192      | 1796 ns     | 1758 ns    | 373333            | 1706 ns    | 1706 ns   | 408267           | 1276 ns      | 1276 ns     | 540025             |
| BM_dot_product/3/8192      | 846 ns      | 854 ns     | 896000            | 842 ns     | 842 ns    | 833770           | 650 ns       | 650 ns      | 1080996            |
| BM_dot_product/4/8192      | 489 ns      | 500 ns     | 1000000           | 529 ns     | 529 ns    | 1301488          | 598 ns       | 598 ns      | 1172999            |
| BM_dot_product/1/16384     | 1796 ns     | 1765 ns    | 407273            | 3420 ns    | 3420 ns   | 204340           | 1757 ns      | 1757 ns     | 389621             |
| BM_dot_product/2/16384     | 3619 ns     | 3530 ns    | 194783            | 3431 ns    | 3431 ns   | 204736           | 2583 ns      | 2583 ns     | 265834             |
| BM_dot_product/3/16384     | 1831 ns     | 1883 ns    | 448000            | 1699 ns    | 1699 ns   | 406700           | 1295 ns      | 1295 ns     | 519284             |
| BM_dot_product/4/16384     | 1228 ns     | 1228 ns    | 560000            | 1089 ns    | 1089 ns   | 629957           | 1172 ns      | 1172 ns     | 591139             |
| BM_cosine_distance/1/1024  | 173 ns      | 171 ns     | 4480000           | 211 ns     | 211 ns    | 3259795          | 167 ns       | 167 ns      | 4177995            |
| BM_cosine_distance/2/1024  | 224 ns      | 225 ns     | 2986667           | 211 ns     | 211 ns    | 3297316          | 218 ns       | 218 ns      | 3262108            |
| BM_cosine_distance/3/1024  | 123 ns      | 120 ns     | 5600000           | 111 ns     | 111 ns    | 6199282          | 114 ns       | 114 ns      | 6073426            |
| BM_cosine_distance/4/1024  | 81.5 ns     | 83.7 ns    | 8960000           | 59.5 ns    | 59.5 ns   | 11877432         | 65.7 ns      | 65.7 ns     | 10613216           |
| BM_cosine_distance/1/2048  | 359 ns      | 353 ns     | 1947826           | 433 ns     | 433 ns    | 1642719          | 332 ns       | 332 ns      | 2118325            |
| BM_cosine_distance/2/2048  | 450 ns      | 439 ns     | 1493333           | 426 ns     | 426 ns    | 1638678          | 432 ns       | 432 ns      | 1606191            |
| BM_cosine_distance/3/2048  | 239 ns      | 229 ns     | 2800000           | 225 ns     | 225 ns    | 3110254          | 229 ns       | 229 ns      | 3046923            |
| BM_cosine_distance/4/2048  | 140 ns      | 136 ns     | 4480000           | 122 ns     | 122 ns    | 5746390          | 125 ns       | 125 ns      | 5432138            |
| BM_cosine_distance/1/4096  | 706 ns      | 680 ns     | 896000            | 851 ns     | 851 ns    | 818869           | 664 ns       | 664 ns      | 1021273            |
| BM_cosine_distance/2/4096  | 917 ns      | 900 ns     | 746667            | 860 ns     | 860 ns    | 812051           | 865 ns       | 865 ns      | 792026             |
| BM_cosine_distance/3/4096  | 458 ns      | 454 ns     | 1445161           | 447 ns     | 447 ns    | 1562938          | 502 ns       | 502 ns      | 1000000            |
| BM_cosine_distance/4/4096  | 265 ns      | 267 ns     | 2635294           | 241 ns     | 241 ns    | 2901303          | 252 ns       | 252 ns      | 2797007            |
| BM_cosine_distance/1/8192  | 1425 ns     | 1395 ns    | 448000            | 1718 ns    | 1718 ns   | 405313           | 1332 ns      | 1332 ns     | 523563             |
| BM_cosine_distance/2/8192  | 1782 ns     | 1758 ns    | 373333            | 1728 ns    | 1728 ns   | 399779           | 1758 ns      | 1758 ns     | 399745             |
| BM_cosine_distance/3/8192  | 935 ns      | 935 ns     | 1120000           | 913 ns     | 913 ns    | 762412           | 920 ns       | 920 ns      | 730501             |
| BM_cosine_distance/4/8192  | 501 ns      | 502 ns     | 1120000           | 580 ns     | 580 ns    | 1208493          | 631 ns       | 631 ns      | 1108676            |
| BM_cosine_distance/1/16384 | 2694 ns     | 2773 ns    | 298667            | 3440 ns    | 3440 ns   | 203346           | 2673 ns      | 2673 ns     | 263766             |
| BM_cosine_distance/2/16384 | 3731 ns     | 3575 ns    | 179200            | 3466 ns    | 3466 ns   | 202783           | 3508 ns      | 3508 ns     | 198097             |
| BM_cosine_distance/3/16384 | 1946 ns     | 2009 ns    | 373333            | 1839 ns    | 1839 ns   | 382622           | 1837 ns      | 1837 ns     | 381716             |
| BM_cosine_distance/4/16384 | 1399 ns     | 1395 ns    | 560000            | 1240 ns    | 1240 ns   | 569119           | 1267 ns      | 1267 ns     | 554124             |
| BM_l1_distance/1/1024      | 128 ns      | 125 ns     | 6400000           | 199 ns     | 199 ns    | 3509361          | 110 ns       | 110 ns      | 6408907            |
| BM_l1_distance/2/1024      | 213 ns      | 213 ns     | 3446154           | 199 ns     | 199 ns    | 3421705          | 151 ns       | 151 ns      | 4542396            |
| BM_l1_distance/3/1024      | 98.7 ns     | 100 ns     | 7466667           | 90.8 ns    | 90.8 ns   | 7434131          | 114 ns       | 114 ns      | 9181592            |
| BM_l1_distance/4/1024      | 52.8 ns     | 53 ns      | 11200000          | 51.9 ns    | 51.9 ns   | 13334502         | 44.4 ns      | 44.4 ns     | 12730004           |
| BM_l1_distance/1/2048      | 231 ns      | 220 ns     | 2986667           | 415 ns     | 415 ns    | 1690429          | 219 ns       | 219 ns      | 2986092            |
| BM_l1_distance/2/2048      | 435 ns      | 429 ns     | 1493333           | 415 ns     | 415 ns    | 1682359          | 319 ns       | 319 ns      | 2170577            |
| BM_l1_distance/3/2048      | 210 ns      | 218 ns     | 3733333           | 205 ns     | 204 ns    | 3376457          | 159 ns       | 159 ns      | 4421009            |
| BM_l1_distance/4/2048      | 112 ns      | 113 ns     | 7466667           | 117 ns     | 117 ns    | 5973547          | 93.4 ns      | 93.4 ns     | 7619189            |
| BM_l1_distance/1/4096      | 462 ns      | 471 ns     | 1493333           | 847 ns     | 847 ns    | 819386           | 440 ns       | 440 ns      | 1610553            |
| BM_l1_distance/2/4096      | 865 ns      | 879 ns     | 640000            | 845 ns     | 845 ns    | 816821           | 646 ns       | 646 ns      | 1070798            |
| BM_l1_distance/3/4096      | 437 ns      | 439 ns     | 1600000           | 419 ns     | 419 ns    | 1673956          | 318 ns       | 318 ns      | 2175732            |
| BM_l1_distance/4/4096      | 229 ns      | 225 ns     | 2986667           | 257 ns     | 257 ns    | 2717743          | 198 ns       | 198 ns      | 3533783            |
| BM_l1_distance/1/8192      | 935 ns      | 942 ns     | 896000            | 1707 ns    | 1707 ns   | 398941           | 875 ns       | 875 ns      | 779632             |
| BM_l1_distance/2/8192      | 1789 ns     | 1779 ns    | 448000            | 1707 ns    | 1707 ns   | 402967           | 1296 ns      | 1296 ns     | 540974             |
| BM_l1_distance/3/8192      | 864 ns      | 858 ns     | 746667            | 850 ns     | 850 ns    | 810368           | 663 ns       | 663 ns      | 1026355            |
| BM_l1_distance/4/8192      | 484 ns      | 487 ns     | 1445161           | 558 ns     | 558 ns    | 1240620          | 605 ns       | 605 ns      | 1152246            |
| BM_l1_distance/1/16384     | 1829 ns     | 1800 ns    | 373333            | 3426 ns    | 3426 ns   | 203558           | 1780 ns      | 1780 ns     | 383330             |
| BM_l1_distance/2/16384     | 3466 ns     | 3530 ns    | 194783            | 3444 ns    | 3444 ns   | 201842           | 2594 ns      | 2594 ns     | 265590             |
| BM_l1_distance/3/16384     | 1908 ns     | 1918 ns    | 407273            | 1711 ns    | 1711 ns   | 407681           | 1324 ns      | 1324 ns     | 529667             |
| BM_l1_distance/4/16384     | 1321 ns     | 1318 ns    | 497778            | 1141 ns    | 1141 ns   | 612057           | 1232 ns      | 1232 ns     | 562264             |
| BM_l1_norm/1/1024          | 108 ns      | 105 ns     | 7466667           | 182 ns     | 182 ns    | 3818439          | 99.9 ns      | 99.9 ns     | 7006081            |
…