@ankane 2023-10-24
Hi @aytekinar, thanks for the PR! It's a neat idea, and the code looks really clean. However, it adds a good amount of complexity.

I think a minimal version would be:

1. AVX only (SSE is always available for x86-64, and AVX is more widely available than AVX-512F)
2. Only if AVX is not enabled by compilation flags
3. Only functions that provide the most benefit (L2 distance squared, inner product, and cosine distance)

For benchmarking, it'd be best use ann-benchmarks to see the impact on the total time.

Overall, I'm not sure it's something I'd like to move forward with right now, but will think on it for a bit (and welcome other comments).