@xfalcox 2023-10-24
> To alleviate the above, I have implemented the SSE, AVX and AVX512F versions of the vector operations, and added a CPU dispatching mechanism (during extension load time) to pick the most recent version the underlying CPU supports (by following the best practices mentioned in Agner Fog's [Optimizing Software in C++ (Chapters 12 and 13)](https://www.agner.org/optimize/#manuals) manual).

That's refreshing to hear. At @discourse we had problems with dependencies shipping AVX improvements, that were followed by the random user running the application in a 13 years old CPU frustrated because it broke.

> Overall, I'm not sure it's something I'd like to move forward with right now, but will think on it for a bit (and welcome other comments).

For us, the CPU cost of the "search" (inner_product) is what is holding us back on enabling this for more of our hosted forums, as most of the 500M+ monthly pageviews we serve will do one search to show related topics when we enable our embeddings feature for a site. We do cache the result of the query for a whole week, but over a million topics are created month again, and with anon visits coming from Google Search, the cache isn't always there.

That all to say that a (17% ~ 50%) speed-up would be *very* welcome for us. I do appreciate the the concern over complexity tho.