@jkatz 2023-10-24
I want to work on benchmarking over the coming days. I do think there's merit to being explicit about SIMD, but I'm not convinced on adding the complexity based on the above data. That said, I do generally think this is a good idea and, at least on paper, supportive of doing this.

The main issues I've seen as of late with the latest version of pgvector is calculations getting CPU bound on high levels of concurrency with higher levels of `hnsw.ef_search`. Specifically, I've observed that HNSW searches scale linearly quite well but they start to fall off with `hnsw.ef_search` that's 200+ with 48+ clients connecting (this is on a dataset of 10MM 128-dim vectors).

That said, the benchmarks above don't convince me yet around adding them -- these are very short tests on very small data sets. I agree with @ankane on running these against ANN Benchmarks (I'm happy to do so), but I also want to test something similar to the proposed `pgbench` method at a much higher concurrency scale over a longer duration. If this continues to show a significant gain, I would be in favor of getting this in sooner.

On the code, I'd recommending keeping the SIMD implementation work to its own file (upstream PostgreSQL does this in a `simd.c`) so that way `vector.c` can just contain the implementation of the vector calculation functions. That way it's easier to maintain/grok the SIMD work.

Thanks!