@ashvardanian 2023-10-26
Hello everyone,

A colleague recently highlighted the potential benefits of integrating SIMD similarity measures into pgvector. To my delight, I discovered this ongoing implementation, which aligns perfectly with a project of mine.

I am the creator of a library named [SimSIMD](https://github.com/ashvardanian/simsimd). It is tailored to implement various similarity measures, including L2, Cosine, Inner Product, Jaccard, Hamming, and Jensen Shannon distances. It uses AVX2 and AVX-512 techniques, including VNNI for int8 and FP16 extensions for half-precision floating points on x86 platforms, including the most recent Sapphire Rapids. Furthermore, it supports Arm NEON and SVE, making it compatible with Graviton 3 and other recent chipsets.

One of the major deployments of SimSIMD is in [USearch](https://github.com/unum-cloud/usearch). This tool has gained traction and is now integrated into platforms like ClickHouse and various data lakes.

I've chronicled some benchmarks in the following articles. While they provide insight into the performance of SimSIMD, I'm keen to determine their relevance to this context:

1. [GCC 12 is 119x slower than AVX-512 SIMD on this task](https://ashvardanian.com/posts/gcc-12-vs-avx512fp16/)
2. [SciPy distances... up to 200x faster with AVX-512 & SVE](https://ashvardanian.com/posts/simsimd-faster-scipy/)

@ankane and @jkatz, please let me know if the integration makes sense. It should be as easy as adding a git submodule and invoking [`simsimd_metric_punned`](https://github.com/ashvardanian/SimSIMD/blob/cdf2c1fb3c84f2a7bd71543eb13d6c0ccb46e9b2/include/simsimd/simsimd.h#L402) to get the appropriate function pointer. I'm happy to help with the implementation and testing.