@ankane 2023-10-26
Hi @ashvardanian, thanks for sharing. Looks really neat, but don't want to take on any external dependencies or different licenses (however, if someone decides to fork and benchmark, I'd be curious to see the results).