@ashvardanian 2023-10-27
@ankane, the license is not an issue, I can dual-license to be compatible, and have previously done the for [StringZilla](https://github.com/ashvardanian/StringZilla), my strings processing libraries. As for benchmarking, shouldn't be hard, just let me know how you generally do that 🤗