@ankane 2024-04-08
Added CPU dispatching for `halfvec` distance functions in the commit above (likely still needs a few tweaks).

@aytekinar, added you as a co-author since your work in this PR was very helpful for this.

The reason for intrinsics / dispatching in this case was a significant difference in performance on x86-64. With the SIFT 1M dataset (128 dimensions):

Benchmark | Before | After
--- | --- | --- 
HNSW build time |  1431 sec | 334 sec
HNSW query time (10k) | 17 sec | 11 sec

Still looking at dispatching for `vector` functions.